@@ -13,7 +13,7 @@ classifiers = ["Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Atmospheric Science"
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.10"
 dynamic = ["version", "readme"]
 dependencies = ["python-dateutil", "numpy", "xarray", "scipy", "pymap3d", "h5py"]
 
